54895/fix algorithm prerequis

See merge request 2021-2022/prjg5/equipe1!7

+++ database/migrations/2021_11_05_234200_create_files_table.php
@@ -18,7 +18,7 @@ class CreateFilesTable extends Migration
         Schema::create('files', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('student');
+            $table->unsignedBigInteger('student');
             $table->foreign('student')->on('student')->references('id')->onDelete('cascade');
 
             // $table->string('type'); // TODO: Type de fichier??
